+++ src/main/java/ladysnake/illuminations/client/IlluminationsClient.java
@@ -1,19 +1,17 @@
 package ladysnake.illuminations.client;
 
 import ladysnake.illuminations.client.renders.entities.FireflyRender;
-import ladysnake.illuminations.client.renders.entities.LightningBugRender;
-import ladysnake.illuminations.client.renders.entities.WillOWispRender;
-import ladysnake.illuminations.common.entities.FireflyEntity;
-import ladysnake.illuminations.common.entities.LightningBugEntity;
-import ladysnake.illuminations.common.entities.WillOWispEntity;
 import ladysnake.illuminations.common.init.IlluminationsBlocks;
+import ladysnake.illuminations.common.init.IlluminationsEntities;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.render.ColorProviderRegistry;
-import net.fabricmc.fabric.api.client.render.EntityRendererRegistry;
+import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityRendererRegistry;
 import net.minecraft.block.Blocks;
 import net.minecraft.client.color.block.BlockColorProvider;
+import net.minecraft.client.render.RenderLayer;
 
 @Environment(EnvType.CLIENT)
 public class IlluminationsClient implements ClientModInitializer {
@@ -22,12 +20,11 @@ public class IlluminationsClient implements ClientModInitializer {
     public void onInitializeClient() {
         registerRenders();
         registerColors();
+        registerCutouts();
     }
 
     public static void registerRenders() {
-        EntityRendererRegistry.INSTANCE.register(FireflyEntity.class, (manager, context) -> new FireflyRender<>(manager));
-        EntityRendererRegistry.INSTANCE.register(LightningBugEntity.class, (manager, context) -> new LightningBugRender<>(manager));
-        EntityRendererRegistry.INSTANCE.register(WillOWispEntity.class, (manager, context) -> new WillOWispRender<>(manager));
+        EntityRendererRegistry.INSTANCE.register(IlluminationsEntities.FIREFLY, (manager, context) -> new FireflyRender<>(manager));
     }
 
     public static void registerColors() {
@@ -39,6 +36,12 @@ public class IlluminationsClient implements ClientModInitializer {
             BlockColorProvider provider = ColorProviderRegistry.BLOCK.get(Blocks.GRASS);
             return provider == null ? -1 : provider.getColor(block, pos, world, layer);
         }, IlluminationsBlocks.FIREFLY_TALL_GRASS);
+    }
+
+    public static void registerCutouts() {
+        BlockRenderLayerMap.INSTANCE.putBlock(IlluminationsBlocks.FIREFLY_GRASS, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(IlluminationsBlocks.FIREFLY_TALL_GRASS, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(IlluminationsBlocks.FIREFLY_IN_A_BOTTLE, RenderLayer.getCutout());
     }
 
 }