+++ src/main/java/ladysnake/illuminations/client/renders/entities/FireflyRender.java
@@ -1,117 +1,107 @@
 package ladysnake.illuminations.client.renders.entities;
 
-import com.mojang.blaze3d.platform.GLX;
-import com.mojang.blaze3d.platform.GlStateManager;
 import ladysnake.illuminations.common.Illuminations;
 import ladysnake.illuminations.common.entities.FireflyEntity;
-import net.minecraft.client.render.BufferBuilder;
-import net.minecraft.client.render.Tessellator;
-import net.minecraft.client.render.VertexFormats;
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.render.OverlayTexture;
+import net.minecraft.client.render.RenderLayer;
+import net.minecraft.client.render.VertexConsumer;
+import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.EntityRenderDispatcher;
 import net.minecraft.client.render.entity.EntityRenderer;
+import net.minecraft.client.util.math.Matrix3f;
+import net.minecraft.client.util.math.Matrix4f;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.client.util.math.Vector3f;
 import net.minecraft.util.Identifier;
-import org.lwjgl.opengl.GL11;
+import net.minecraft.util.math.MathHelper;
 
-import javax.annotation.Nonnull;
-import java.util.Random;
+public class FireflyRender<T extends FireflyEntity> extends EntityRenderer<T> {
+    private static final Identifier TEXTURE = new Identifier(Illuminations.MOD_ID, "textures/entity/firefly.png");
+    private static final Identifier OVERLAY = new Identifier(Illuminations.MOD_ID, "textures/entity/firefly_overlay.png");
+    private static final RenderLayer TEXTURE_LAYER = RenderLayer.getEntityTranslucent(TEXTURE);
+    private static final RenderLayer OVERLAY_LAYER = RenderLayer.getEntityTranslucent(OVERLAY);
 
-
-public class FireflyRender<E extends FireflyEntity> extends EntityRenderer<E> {
-    public FireflyRender(EntityRenderDispatcher renderManager) {
-        super(renderManager);
-        this.field_4672 = 0;
+    public FireflyRender(EntityRenderDispatcher entityRenderDispatcher) {
+        super(entityRenderDispatcher);
+        this.shadowSize = 0F;
+        this.shadowDarkness = 0F;
     }
 
-    @Override
-    public void render(@Nonnull E entity, double x, double y, double z, float entityYaw, float tickDelta) {
-        if (!this.renderOutlines) {
-            GlStateManager.pushMatrix();
+    protected int getBlockLight(T firefly, float f) {
+        return MathHelper.clamp(super.getBlockLight(firefly, f) + 7, 0, 15);
+    }
 
-            GlStateManager.translatef((float) x, (float) y + 0.1f, (float) z);
+    public void render(T firefly, float f, float g, MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i) {
+        matrixStack.push();
+        float ux = 0;
+        float uy = 1;
+        float vx = 0;
+        float vy = 1;
+        int red = (int) (firefly.getColorModifier() * 255);
+        float scale = firefly.getScaleFactor();
+        matrixStack.translate(0, 0, 0);
+        matrixStack.multiply(this.renderManager.getRotation());
+        matrixStack.multiply(Vector3f.POSITIVE_Y.getDegreesQuaternion(180.0F));
+        matrixStack.scale(scale, scale, scale);
 
-            GlStateManager.enableBlend();
-            GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
-            GlStateManager.disableLighting();
+        int alpha = firefly.getAlpha();
+        int nextAlphaGoal = firefly.getNextAlphaGoal();
 
-            GLX.glMultiTexCoord2f(GLX.GL_TEXTURE1, 240f, 240f);
-
-            GlStateManager.rotatef(180.0F - this.renderManager.cameraYaw, 0.0F, 1.0F, 0.0F);
-            GlStateManager.rotatef((float) (this.renderManager.gameOptions.perspective == 2 ? -1 : 1) * -this.renderManager.cameraPitch, 1.0F, 0.0F, 0.0F);
-
-            this.bindEntityTexture(entity);
-            float alpha = entity.getAlpha();
-            float scale = entity.getScaleModifier();
-            float color = entity.getColorModifier();
-            Float nextAlphaGoal = entity.getNextAlphaGoal();
-
-            // if just spawned
-            if (entity.age < 50) {
-                alpha = 0;
-            }
-
+        // if just spawned
+        if (firefly.age < 50) {
+            alpha = 0;
+        } else {
             // if day
-            float tod = entity.world.getLevelProperties().getTimeOfDay();
+            float tod = firefly.world.getLevelProperties().getTimeOfDay() % 24000;
             if (tod >= 1000 && tod < 13000) {
-                nextAlphaGoal = 0.0F;
+                nextAlphaGoal = 0;
             }
 
-            if (nextAlphaGoal == null || nextAlphaGoal.equals(round(alpha, 1))) {
-                entity.setNextAlphaGoal(new Random().nextInt(11) / 10.0F);
+            if (alpha > nextAlphaGoal - 10 && alpha < nextAlphaGoal + 10) {
+                firefly.setNextAlphaGoal(firefly.getRandom().nextInt(256));
             } else {
                 if (nextAlphaGoal > alpha) {
-                    alpha += 0.05F;
+                    alpha += 10;
                 } else if (nextAlphaGoal < alpha) {
-                    alpha -= 0.05F;
+                    alpha -= 10;
                 }
             }
-
-            entity.setAlpha(Math.min(Math.max(alpha, 0), 1));
-            GlStateManager.scalef(scale, scale, scale);
-            GlStateManager.color4f(color, 1F, 0F, entity.getAlpha());
-
-            Tessellator tessellator = Tessellator.getInstance();
-            BufferBuilder bufferbuilder = tessellator.getBufferBuilder();
-            float minU = 0;
-            float minV = 0;
-            float maxU = 1;
-            float maxV = 1;
-            bufferbuilder.begin(GL11.GL_QUADS, VertexFormats.POSITION_UV_NORMAL);
-            bufferbuilder.vertex(-0.5D, -0.25D, 0.0D).texture((double) maxU, (double) maxV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(0.5D, -0.25D, 0.0D).texture((double) minU, (double) maxV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(0.5D, 0.75D, 0.0D).texture((double) minU, (double) minV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(-0.5D, 0.75D, 0.0D).texture((double) maxU, (double) minV).normal(0.0F, 1.0F, 0.0F).next();
-            tessellator.draw();
-
-            this.bindTexture(new Identifier(Illuminations.MOD_ID, "textures/entity/firefly_overlay.png"));
-            GlStateManager.color4f(1F, 1F, 1F, entity.getAlpha());
-            bufferbuilder.begin(7, VertexFormats.POSITION_UV_NORMAL);
-            bufferbuilder.vertex(-0.5D, -0.25D, 0.0D).texture((double) maxU, (double) maxV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(0.5D, -0.25D, 0.0D).texture((double) minU, (double) maxV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(0.5D, 0.75D, 0.0D).texture((double) minU, (double) minV).normal(0.0F, 1.0F, 0.0F).next();
-            bufferbuilder.vertex(-0.5D, 0.75D, 0.0D).texture((double) maxU, (double) minV).normal(0.0F, 1.0F, 0.0F).next();
-            tessellator.draw();
-
-            GlStateManager.disableBlend();
-            GlStateManager.disableRescaleNormal();
-            GlStateManager.enableLighting();
-            GlStateManager.popMatrix();
-            super.render(entity, x, y, z, entityYaw, tickDelta);
         }
+
+        if (MinecraftClient.getInstance().isPaused()) {
+            alpha = firefly.getAlpha();
+        }
+        firefly.setAlpha(Math.min(Math.max(alpha, 0), 255));
+        alpha = firefly.getAlpha();
+
+        // firefly
+        VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(TEXTURE_LAYER);
+        MatrixStack.Entry entry = matrixStack.peek();
+        Matrix4f matrix4f = entry.getModel();
+        Matrix3f matrix3f = entry.getNormal();
+        method_23171(vertexConsumer, matrix4f, matrix3f, -0.5F, -0.25F, red, 0, alpha, ux, vy);
+        method_23171(vertexConsumer, matrix4f, matrix3f, 0.5F, -0.25F, red, 0, alpha, uy, vy);
+        method_23171(vertexConsumer, matrix4f, matrix3f, 0.5F, 0.75F, red, 0, alpha, uy, vx);
+        method_23171(vertexConsumer, matrix4f, matrix3f, -0.5F, 0.75F, red, 0, alpha, ux, vx);
+
+        // firefly overlay
+        vertexConsumer = vertexConsumerProvider.getBuffer(OVERLAY_LAYER);
+        method_23171(vertexConsumer, matrix4f, matrix3f, -0.5F, -0.25F, 255, 255, alpha, ux, vy);
+        method_23171(vertexConsumer, matrix4f, matrix3f, 0.5F, -0.25F, 255, 255, alpha, uy, vy);
+        method_23171(vertexConsumer, matrix4f, matrix3f, 0.5F, 0.75F, 255, 255, alpha, uy, vx);
+        method_23171(vertexConsumer, matrix4f, matrix3f, -0.5F, 0.75F, 255, 255, alpha, ux, vx);
+
+        matrixStack.pop();
+        super.render(firefly, f, g, matrixStack, vertexConsumerProvider, i);
     }
 
-//    @Override
-//    public void renderShadow(Entity entityIn, double x, double y, double z, float yaw, float partialTicks) {
-//    }
-
-    @Override
-    protected Identifier getTexture( E entity) {
-        return new Identifier(Illuminations.MOD_ID, "textures/entity/firefly.png");
+    private static void method_23171(VertexConsumer vertexConsumer, Matrix4f matrix4f, Matrix3f matrix3f, float x, float y, int red, int blue, int alpha, float u, float v) {
+        vertexConsumer.vertex(matrix4f, x, y, 0.0F).color(red, 255, blue, alpha).texture(u, v).overlay(OverlayTexture.DEFAULT_UV).light(15728880).normal(matrix3f, 0.0F, 1.0F, 0.0F).next();
     }
 
-    // Useful method
-    private static float round(float value, int precision) {
-        int scale = (int) Math.pow(10, precision);
-        return (float) Math.round(value * scale) / scale;
+    public Identifier getTexture(T firefly) {
+        return TEXTURE;
     }
 
 }