+++ src/main/java/ladysnake/illuminations/common/init/IlluminationsBlocks.java
@@ -17,13 +17,12 @@ public class IlluminationsBlocks {
     public static Block FIREFLY_GRASS;
     public static Block FIREFLY_TALL_GRASS;
 
-
     public static void init() {
-//        GLOWWORM = registerBlock(new GlowwormBlock(FabricBlockSettings.of(Material.REPLACEABLE_PLANT).noCollision().breakInstantly().sounds(BlockSoundGroup.SLIME).lightLevel(5).build()), "glowworm");
-        FIREFLY_IN_A_BOTTLE = registerBlock(new LanternBlock(FabricBlockSettings.of(Material.GLASS).strength(0.3F, 0.3F).sounds(BlockSoundGroup.GLASS).lightLevel(10).build()), "firefly_in_a_bottle");
+        FIREFLY_IN_A_BOTTLE = registerBlock(new LanternBlock(FabricBlockSettings.of(Material.GLASS).strength(0.3F, 0.3F).sounds(BlockSoundGroup.GLASS).lightLevel(15).build()), "firefly_in_a_bottle");
         FIREFLY_GRASS = registerBlock(new FireflyGrassBlock(FabricBlockSettings.of(Material.REPLACEABLE_PLANT).noCollision().breakInstantly().sounds(BlockSoundGroup.GRASS).lightLevel(1).build()), "firefly_grass");
         FIREFLY_TALL_GRASS = registerBlock(new FireflyTallGrassBlock(FabricBlockSettings.of(Material.REPLACEABLE_PLANT).noCollision().breakInstantly().sounds(BlockSoundGroup.GRASS).lightLevel(1).build()), "firefly_tall_grass");
-        // registering flammability
+
+        // registering firefly grass flammability
         ((FireBlock)Blocks.FIRE).registerFlammableBlock(FIREFLY_GRASS, 60, 100);
         ((FireBlock)Blocks.FIRE).registerFlammableBlock(FIREFLY_TALL_GRASS, 60, 100);
     }
@@ -32,12 +31,12 @@ public class IlluminationsBlocks {
         return registerBlock(block, name, true);
     }
 
-    private static Block registerBlock(Block block, String name, boolean doItem) {
+    private static Block registerBlock(Block    block, String name, boolean doItem) {
         Registry.register(Registry.BLOCK, Illuminations.MOD_ID + ":" + name, block);
 
         if (doItem) {
-            BlockItem item = new BlockItem(block, new Item.Settings().itemGroup(ItemGroup.DECORATIONS));
-            item.registerBlockItemMap(Item.BLOCK_ITEM_MAP, item);
+            BlockItem item = new BlockItem(block, new Item.Settings().group(ItemGroup.DECORATIONS));
+            item.appendBlocks(Item.BLOCK_ITEMS, item);
             IlluminationsItems.registerItem(item, name);
         }