+++ src/main/java/ladysnake/illuminations/common/init/IlluminationsItems.java
@@ -3,21 +3,21 @@ package ladysnake.illuminations.common.init;
 import ladysnake.illuminations.common.Illuminations;
 import ladysnake.illuminations.common.items.BugNetItem;
 import ladysnake.illuminations.common.items.FireflyItem;
-import ladysnake.illuminations.common.items.WillOWispItem;
+import ladysnake.illuminations.common.items.GlowMealItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.util.registry.Registry;
 
 public class IlluminationsItems {
 
-    public static Item WILL_O_WISP;
     public static Item BUG_NET;
     public static Item FIREFLY;
+    public static Item GLOW_MEAL;
 
     public static void init() {
-        WILL_O_WISP = registerItem(new WillOWispItem((new Item.Settings())), "will_o_wisp");
-        BUG_NET = registerItem(new BugNetItem((new Item.Settings()).durability(238).itemGroup(ItemGroup.TOOLS)), "bug_net");
-        FIREFLY = registerItem(new FireflyItem((new Item.Settings()).itemGroup(ItemGroup.MISC)), "firefly");
+        BUG_NET = registerItem(new BugNetItem((new Item.Settings()).maxDamage(238).group(ItemGroup.TOOLS)), "bug_net");
+        FIREFLY = registerItem(new FireflyItem((new Item.Settings()).group(ItemGroup.MISC)), "firefly");
+        GLOW_MEAL = registerItem(new GlowMealItem((new Item.Settings()).group(ItemGroup.MISC)), "glow_meal");
     }
 
     public static Item registerItem(Item item, String name) {